OpenConceptLab/ocl_issues#2576 | extracting common component | using translations
