+++ src/components/repos/SourceVersionsTab.jsx
@@ -65,6 +65,7 @@ import {
 } from '../../common/utils';
 import { OperationsContext } from '../app/LayoutContext';
 import AccessIcon from '../common/AccessIcon';
+import MarkdownContent from '../common/MarkdownContent';
 import ConceptIcon from '../concepts/ConceptIcon';
 import MappingIcon from '../mappings/MappingIcon';
 
@@ -111,128 +112,6 @@ const downloadBlob = (response, fallbackName) => {
   link.click();
   link.remove();
   setTimeout(() => window.URL.revokeObjectURL(url), 1000);
-};
-
-const renderInlineMarkdown = text => {
-  const tokens = [];
-  const pattern = /(\*\*([^*]+)\*\*|\*([^*]+)\*|`([^`]+)`|\[([^\]]+)\]\(([^)]+)\))/g;
-  let lastIndex = 0;
-  let match;
-
-  while((match = pattern.exec(text || ''))) {
-    if(match.index > lastIndex)
-      tokens.push(text.slice(lastIndex, match.index));
-
-    if(match[2])
-      tokens.push(<strong key={tokens.length}>{match[2]}</strong>);
-    else if(match[3])
-      tokens.push(<em key={tokens.length}>{match[3]}</em>);
-    else if(match[4])
-      tokens.push(<Box component="code" key={tokens.length} sx={{ fontFamily: 'monospace', px: 0.5, bgcolor: 'surface.main', borderRadius: '3px' }}>{match[4]}</Box>);
-    else if(match[5])
-      tokens.push(<a key={tokens.length} href={match[6]} target="_blank" rel="noopener noreferrer">{match[5]}</a>);
-
-    lastIndex = pattern.lastIndex;
-  }
-
-  if(lastIndex < (text || '').length)
-    tokens.push(text.slice(lastIndex));
-
-  return tokens;
-};
-
-const parseMarkdownTable = lines => {
-  if(lines.length < 2 || !lines[0].trim().startsWith('|') || !lines[1].includes('---'))
-    return null;
-
-  const parseRow = line => line.trim().replace(/^\|/, '').replace(/\|$/, '').split('|').map(cell => cell.trim());
-  const headers = parseRow(lines[0]);
-  const rows = [];
-  let nextIndex = 2;
-
-  while(nextIndex < lines.length && lines[nextIndex].trim().startsWith('|')) {
-    rows.push(parseRow(lines[nextIndex]));
-    nextIndex += 1;
-  }
-
-  return { headers, rows, nextIndex };
-};
-
-const MarkdownContent = ({ markdown }) => {
-  const lines = (markdown || '').split('\n');
-  const elements = [];
-  let index = 0;
-
-  while(index < lines.length) {
-    const line = lines[index];
-    const trimmed = line.trim();
-
-    if(!trimmed) {
-      index += 1;
-      continue;
-    }
-
-    if(trimmed === '---') {
-      elements.push(<Divider key={elements.length} sx={{ my: 2 }} />);
-      index += 1;
-      continue;
-    }
-
-    const table = parseMarkdownTable(lines.slice(index));
-    if(table) {
-      elements.push(
-        <TableContainer key={elements.length} sx={{ my: 2, border: '1px solid', borderColor: 'surface.nv80', borderRadius: '4px' }}>
-          <Table size="small">
-            <TableHead>
-              <TableRow>
-                {table.headers.map(header => <TableCell key={header} sx={{ fontWeight: 700 }}>{renderInlineMarkdown(header)}</TableCell>)}
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {table.rows.map((row, rowIndex) => (
-                <TableRow key={rowIndex}>
-                  {row.map((cell, cellIndex) => <TableCell key={`${rowIndex}-${cellIndex}`}>{renderInlineMarkdown(cell)}</TableCell>)}
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      );
-      index += table.nextIndex;
-      continue;
-    }
-
-    const heading = trimmed.match(/^(#{1,6})\s+(.+)$/);
-    if(heading) {
-      const variant = heading[1].length <= 1 ? 'h6' : 'subtitle1';
-      elements.push(
-        <Typography key={elements.length} variant={variant} sx={{ mt: elements.length ? 2.5 : 0, mb: 1, fontWeight: 700 }}>
-          {renderInlineMarkdown(heading[2])}
-        </Typography>
-      );
-      index += 1;
-      continue;
-    }
-
-    if(trimmed.startsWith('>')) {
-      elements.push(
-        <Box key={elements.length} sx={{ borderLeft: '3px solid', borderColor: 'primary.main', pl: 1.5, my: 1.5, color: 'text.secondary' }}>
-          <Typography variant="body2">{renderInlineMarkdown(trimmed.replace(/^>\s?/, ''))}</Typography>
-        </Box>
-      );
-      index += 1;
-      continue;
-    }
-
-    elements.push(
-      <Typography key={elements.length} variant="body2" sx={{ my: 1, lineHeight: 1.6 }}>
-        {renderInlineMarkdown(trimmed)}
-      </Typography>
-    );
-    index += 1;
-  }
-
-  return <Box sx={{ '& a': { color: 'primary.main' } }}>{elements}</Box>;
 };
 
 const VersionExportDialog = ({ version, open, onClose }) => {
@@ -501,9 +380,9 @@ const SourceVersionsTab = ({
       if(response.detail || response.error)
         setAlert({ severity: 'error', message: formatError(response.detail || response.error) });
       else if(response.status === 202)
-        setAlert({ severity: 'success', message: 'Summary request queued.' });
+        setAlert({ severity: 'success', message: t('repo.summary_request_queued') });
       else
-        setAlert({ severity: 'warning', message: 'Summary request submitted.' });
+        setAlert({ severity: 'warning', message: t('repo.summary_request_submitted') });
     });
   };
   const onExternalExportsChange = updatedVersion => {
@@ -601,7 +480,7 @@ const SourceVersionsTab = ({
                       </Stack>
                     )}
                     {!isHEAD && !version.released && (
-                      <Typography variant="body2">Unreleased</Typography>
+                      <Typography variant="body2">{t('repo.unreleased')}</Typography>
                     )}
                     {version.retired && <Typography variant="caption" color="error" sx={{ display: 'block' }}>{t('common.retired')}</Typography>}
                   </TableCell>
@@ -622,7 +501,7 @@ const SourceVersionsTab = ({
             {!loading && sortedVersions.length === 0 && (
               <TableRow>
                 <TableCell colSpan={6} sx={bodyCellSx}>
-                  <Alert severity="info">No versions found.</Alert>
+                  <Alert severity="info">{t('repo.no_versions_found')}</Alert>
                 </TableCell>
               </TableRow>
             )}
@@ -630,17 +509,27 @@ const SourceVersionsTab = ({
         </Table>
       </TableContainer>
       <Menu anchorEl={menuState.anchorEl} open={Boolean(menuState.anchorEl)} onClose={closeMenu}>
-        <MenuItem onClick={() => withClose(version => onVersionChange(version))}><VisibilityIcon fontSize="small" sx={{ mr: 1 }} />Explore Version</MenuItem>
-        <MenuItem onClick={() => withClose(version => copyVersionURL(version))}><CopyIcon fontSize="small" sx={{ mr: 1 }} />Copy URL</MenuItem>
-        <MenuItem onClick={() => withClose(version => setExportVersion(version))} disabled={!isLoggedIn()}><ExportIcon fontSize="small" sx={{ mr: 1 }} />Export Version</MenuItem>
-        <MenuItem onClick={() => withClose(version => setExternalExportsVersion(version))} disabled={!isLoggedIn() || isHeadVersion(menuState.version)}><ExternalExportIcon fontSize="small" sx={{ mr: 1 }} />External Exports</MenuItem>
-        <MenuItem onClick={() => withClose(version => setChangelogVersion(version))} disabled={!getPreviousVersionURL(menuState.version)}><ChangelogIcon fontSize="small" sx={{ mr: 1 }} />Changelog</MenuItem>
-        <MenuItem onClick={() => withClose(compareVersion)} disabled={!getPreviousVersionURL(menuState.version)}><OpenInNewIcon fontSize="small" sx={{ mr: 1 }} />Compare with Previous</MenuItem>
+        <MenuItem onClick={() => withClose(version => onVersionChange(version))}><VisibilityIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.explore_version')}</MenuItem>
+        <MenuItem onClick={() => withClose(version => copyVersionURL(version))}><CopyIcon fontSize="small" sx={{ mr: 1 }} />{t('common.copy_api_url')}</MenuItem>
+        <MenuItem onClick={() => withClose(version => setExportVersion(version))} disabled={!isLoggedIn()}><ExportIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.export_version')}</MenuItem>
+        <MenuItem onClick={() => withClose(version => setExternalExportsVersion(version))} disabled={!isLoggedIn() || isHeadVersion(menuState.version)}><ExternalExportIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.external_exports')}</MenuItem>
+        <MenuItem onClick={() => withClose(version => setChangelogVersion(version))} disabled={!getPreviousVersionURL(menuState.version)}><ChangelogIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.changelog')}</MenuItem>
+        <MenuItem onClick={() => withClose(compareVersion)} disabled={!getPreviousVersionURL(menuState.version)}><OpenInNewIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.compare_with_previous')}</MenuItem>
         {hasAccess && <Divider />}
         {hasAccess && <MenuItem onClick={() => withClose(onEditVersion)} disabled={isHeadVersion(menuState.version)}><EditIcon fontSize="small" sx={{ mr: 1 }} />{t('common.edit')}</MenuItem>}
-        {hasAccess && <MenuItem onClick={() => withClose(onReleaseVersion)} disabled={isHeadVersion(menuState.version)}><ReleaseIcon fontSize="small" sx={{ mr: 1 }} />{menuState.version?.released ? 'Unrelease Version' : 'Release Version'}</MenuItem>}
-        {hasAccess && <MenuItem onClick={() => withClose(computeSummary)}><SummaryIcon fontSize="small" sx={{ mr: 1 }} />Re-compute Summary</MenuItem>}
-        {hasAccess && <MenuItem onClick={() => withClose(onDeleteVersion)} disabled={menuState.version?.retired}><DeleteIcon fontSize="small" sx={{ mr: 1 }} />Delete</MenuItem>}
+        {hasAccess && <MenuItem onClick={() => withClose(onReleaseVersion)} disabled={isHeadVersion(menuState.version)}><ReleaseIcon fontSize="small" sx={{ mr: 1 }} />{menuState.version?.released ? t('repo.unrelease_version') : t('repo.release_version')}</MenuItem>}
+        {hasAccess && <MenuItem onClick={() => withClose(computeSummary)}><SummaryIcon fontSize="small" sx={{ mr: 1 }} />{t('repo.recompute_summary')}</MenuItem>}
+        {hasAccess && <Divider />}
+        {hasAccess && (
+          <MenuItem
+            onClick={() => withClose(onDeleteVersion)}
+            disabled={menuState.version?.retired}
+            sx={{ color: 'error.main', '& .MuiSvgIcon-root': { color: 'error.main' } }}
+          >
+            <DeleteIcon fontSize="small" sx={{ mr: 1 }} />
+            {t('repo.delete_repo_version')}
+          </MenuItem>
+        )}
       </Menu>
       {Boolean(exportVersion) && <VersionExportDialog version={exportVersion} open={Boolean(exportVersion)} onClose={() => setExportVersion(null)} />}
       {Boolean(externalExportsVersion) && <ExternalExportsDialog version={externalExportsVersion} canEdit={hasAccess} open={Boolean(externalExportsVersion)} onClose={() => setExternalExportsVersion(null)} onChange={onExternalExportsChange} />}
