+++ src/i18n/locales/en/translations.json
@@ -445,6 +445,16 @@
     "error_version_update": "Could not update version",
     "versions_expansions": "Versions + Expansions",
     "versions": "Versions",
+    "unreleased": "Unreleased",
+    "explore_version": "Explore Version",
+    "export_version": "Export Version",
+    "external_exports": "External Exports",
+    "changelog": "Changelog",
+    "compare_with_previous": "Compare with Previous",
+    "recompute_summary": "Re-compute Summary",
+    "summary_request_queued": "Summary request queued.",
+    "summary_request_submitted": "Summary request submitted.",
+    "no_versions_found": "No versions found.",
     "expansions": "Expansions",
     "versions_expansions_subtitle": "HEAD, released versions, and their expansions",
     "autoexpand": "Auto-expand",
