+++ src/i18n/locales/zh/translations.json
@@ -217,6 +217,8 @@
     "add_concept": "添加概念",
     "source": "来源",
     "release_status": "发行版本状态",
+    "release_version": "Release Version",
+    "unrelease_version": "Un-Release Version",
     "compare_versions": "比较版本",
     "compare": "比较",
     "visibility": "可见性",
@@ -250,6 +252,18 @@
     "summary_concepts_references": "{{count}} 条概念引用关系",
     "summary_mappings_references": "{{count}} 条映射关系引用关系",
     "refresh_summary": "刷新摘要",
+    "versions": "版本",
+    "unreleased": "Unreleased",
+    "explore_version": "Explore Version",
+    "export_version": "Export Version",
+    "external_exports": "External Exports",
+    "changelog": "Changelog",
+    "compare_with_previous": "Compare with Previous",
+    "recompute_summary": "Re-compute Summary",
+    "summary_request_queued": "Summary request queued.",
+    "summary_request_submitted": "Summary request submitted.",
+    "no_versions_found": "No versions found.",
+    "delete_repo_version": "Delete Repo Version",
     "repo_version_summary": "存储库版本摘要",
     "repo_summary_is_calculating": "正在计算处理存储库摘要数据，请请稍后再回来查看。",
     "no_overview": "所有者尚未为该存储库创建概述页面"
